+++ CHANGELOG.rst
@@ -3,8 +3,8 @@ Change Log
 
 Development Version:
 
-- Add chunking method ``chunks="h5netcdf"`` with better handling of
-  unlimited dimensions.
+- Add ``chunking_heuristic`` keyword and custom heuristic ``chunking_heuristic="h5netcdf"``
+  with better handling of unlimited dimensions.
   By `Dion Häfner <https://github.com/dionhaefner>`_.
 - Return group name instead of full group path for legacy API.
   By `Kai Mühlbauer <https://github.com/kmuehlbauer>`_.